refactor(workspace,fuji): split markdown into an editable vault and a read-only export
"Markdown file" was two products wearing one name. An editable directory and a
read-only projection have opposite requirements: the editable path needs a rigid
convention so the reconcile is safe, the read-only path wants free serialization.
Forcing both through one configurable seam is what produced the RoundTripUnproven
runtime guard (a check compensating for a config that could not prove its own
round trip).

Split them:

- attachMarkdownVault: the editable, two-way seam. Rigid by design:
  <dir>/<table>/<id>.md, frontmatter IS the row, so the round trip is identity
  and cannot lose data (the guard is gone, not replaced). Continuous materialize
  plus markdown_apply (disk to Yjs, guarded, atomic) and markdown_rebuild. An
  optional per-table readBody adds a read-only body section.
- attachMarkdownExport: the read-only projection. Continuous Yjs to disk only,
  free serialization (filename, toMarkdown, per-table dir), no import path. The
  sqlite materializer is the read-only sibling for a relational projection.

Both share one observe-driven materialize core (shared.ts) parameterized by the
single thing they disagree on: how a row renders to a file. Files are now always
<id>.md (slugFilename and the custom filename option are removed): the id is the
stable handle a reconcile keys on and an internal link targets, and a clean break
has no data to migrate. fuji, tab-manager, and honeycrisp move to the vault; fuji
materializes its body read-only via readBody and soft-deletes via onDelete.
Bodies are read-only output in v1; body import is gated on a faithful codec.
