+++ apps/honeycrisp/project.ts
@@ -11,9 +11,8 @@ import { EPICENTER_API_URL } from '@epicenter/constants/apps';
 import { defineActions, defineWorkspace } from '@epicenter/workspace';
 import { defineMount } from '@epicenter/workspace/daemon';
 import {
-	attachMarkdownMaterializer,
+	attachMarkdownVault,
 	type GitAutosaveConfig,
-	slugFilename,
 } from '@epicenter/workspace/document/materializer/markdown';
 import { attachBunSqliteMaterializer } from '@epicenter/workspace/document/materializer/sqlite';
 import {
@@ -61,9 +60,9 @@ export function honeycrisp(opts: HoneycrispMountOptions = {}) {
 				log: createLogger(`${mount}-sqlite`),
 			});
 
-			const markdown = attachMarkdownMaterializer(workspace, {
+			const markdown = attachMarkdownVault(workspace, {
 				dir: mdDir,
-				perTable: { notes: { filename: slugFilename('title') } },
+				tables: { notes: {} },
 				git: opts.git,
 			});
