+++ apps/tab-manager/project.ts
@@ -10,9 +10,8 @@ import { EPICENTER_API_URL } from '@epicenter/constants/apps';
 import { defineActions, defineWorkspace } from '@epicenter/workspace';
 import { defineMount } from '@epicenter/workspace/daemon';
 import {
-	attachMarkdownMaterializer,
+	attachMarkdownVault,
 	type GitAutosaveConfig,
-	slugFilename,
 } from '@epicenter/workspace/document/materializer/markdown';
 import { attachBunSqliteMaterializer } from '@epicenter/workspace/document/materializer/sqlite';
 import {
@@ -66,12 +65,12 @@ export function tabManager(opts: TabManagerMountOptions = {}) {
 				},
 				log: createLogger(`${mount}-sqlite`),
 			});
-			const markdown = attachMarkdownMaterializer(workspace, {
+			const markdown = attachMarkdownVault(workspace, {
 				dir: mdDir,
-				perTable: {
-					bookmarks: { filename: slugFilename('title') },
+				tables: {
+					bookmarks: {},
 					devices: {},
-					savedTabs: { filename: slugFilename('title') },
+					savedTabs: {},
 				},
 				git: opts.git,
 			});
